[FIX] ERD 수정
