+++ src/main/java/com/soptie/server/doll/entity/DollImage.java
@@ -1,0 +1,11 @@
+package com.soptie.server.doll.entity;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public class DollImage {
+
+	private String faceImageUrl;
+
+	private String attentionImageUrl;
+}
