+++ src/main/java/com/soptie/server/memberRoutine/entity/daily/CompletedMemberDailyRoutine.java
@@ -1,6 +1,4 @@
-package com.soptie.server.memberRoutine.entity;
-
-import java.util.Objects;
+package com.soptie.server.memberRoutine.entity.daily;
 
 import com.soptie.server.member.entity.Member;
 import com.soptie.server.routine.entity.daily.DailyRoutine;
@@ -17,7 +15,7 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor
-public class MemberDailyRoutine {
+public class CompletedMemberDailyRoutine {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -26,8 +24,6 @@ public class MemberDailyRoutine {
 
 	private int achieveCount;
 
-	private boolean isAchieve;
-
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "member_id")
 	private Member member;
@@ -35,19 +31,4 @@ public class MemberDailyRoutine {
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "routine_id")
 	private DailyRoutine routine;
-
-	public MemberDailyRoutine(Member member, DailyRoutine routine) {
-		this.achieveCount = 0;
-		this.isAchieve = false;
-		setMember(member);
-		this.routine = routine;
-	}
-
-	private void setMember(Member member) {
-		if (Objects.nonNull(this.member)) {
-			this.member.getDailyRoutines().remove(this);
-		}
-		this.member = member;
-		member.getDailyRoutines().add(this);
-	}
 }
