+++ src/main/java/com/soptie/server/memberRoutine/entity/daily/MemberDailyRoutine.java
@@ -1,4 +1,6 @@
-package com.soptie.server.memberRoutine.entity;
+package com.soptie.server.memberRoutine.entity.daily;
+
+import java.util.Objects;
 
 import com.soptie.server.member.entity.Member;
 import com.soptie.server.routine.entity.daily.DailyRoutine;
@@ -15,7 +17,7 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor
-public class CompletedMemberDailyRoutine {
+public class MemberDailyRoutine {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -24,6 +26,8 @@ public class CompletedMemberDailyRoutine {
 
 	private int achieveCount;
 
+	private boolean isAchieve;
+
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "member_id")
 	private Member member;
@@ -31,4 +35,19 @@ public class CompletedMemberDailyRoutine {
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "routine_id")
 	private DailyRoutine routine;
+
+	public MemberDailyRoutine(Member member, DailyRoutine routine) {
+		this.achieveCount = 0;
+		this.isAchieve = false;
+		setMember(member);
+		this.routine = routine;
+	}
+
+	private void setMember(Member member) {
+		if (Objects.nonNull(this.member)) {
+			this.member.getDailyRoutines().remove(this);
+		}
+		this.member = member;
+		member.getDailyRoutines().add(this);
+	}
 }
