+++ src/main/java/com/soptie/server/routine/entity/daily/RoutineImage.java
@@ -1,0 +1,12 @@
+package com.soptie.server.routine.entity.daily;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public class RoutineImage {
+
+	private String coloredIconImageUrl;
+	private String uncoloredIconImageUrl;
+
+	private String backgroundImageUrl;
+}
