+++ src/main/java/com/soptie/server/routine/entity/happiness/RoutineImage.java
@@ -1,0 +1,11 @@
+package com.soptie.server.routine.entity.happiness;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public class RoutineImage {
+
+	private String iconImageUrl;
+
+	private String contentImageUrl;
+}
